requirements.txt
chandra-ocr>=0.1.0
pandas>=2.0.0
openpyxl>=3.1.0
psutil>=5.9.0
pdf2image>=1.16.3
Pillow>=10.0.0
easyocr>=1.7.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

